routers: units: return full PSU objects
This change is to allow monitoring applications to see all PSU units on
an instance in a single call. The PSU and channels are also flattened in
a list so that json_exporter can access data from the parent (in this
case the PSU).

Other options (such as data-exporter) are available, but this seems to
be the most direct path.

Signed-off-by: Liam Beguin <liambeguin@gmail.com>