routes: switch to serializer classes
It proves a bit messy and hard to have a common.channel/psu abstract
class that can be serialized properly. Make sure that all derived
classes of common.channel/psu can be serialized by the
types.{PSU,Channel}Models. This might be manual to keep in sync but
we're not very likely to add many more parameters.

Signed-off-by: Liam Beguin <liambeguin@gmail.com>